requirements.txt
flask==3.1.0
mysql-connector-python==9.1.0
ttkbootstrap== 1.12.0
pillow==11.1.0
matplotlib==3.10.1
customtkinter== 5.2.2